Refactor: Rename handler to workflow_tasks
